requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests